my-python-web3-project/requirements.txt
web3==7.13.0
PyYAML==6.0.2